+++ frontend/app/src/entities/resource-manager/domain/type.ts
@@ -1,0 +1,9 @@
+export type NumberPool = {
+  id: string;
+  label: string;
+  kind: string;
+  nodeAttribute: {
+    id: string;
+    name: string;
+  };
+};
